+++ src/Database/Esqueleto/Experimental.hs
@@ -65,6 +65,7 @@ module Database.Esqueleto.Experimental
     -- * The Normal Stuff
     , where_
     , groupBy
+    , groupBy_
     , orderBy
     , rand
     , asc
@@ -85,6 +86,7 @@ module Database.Esqueleto.Experimental
 
     , val
     , isNothing
+    , isNothing_
     , just
     , nothing
     , joinV
@@ -174,6 +176,7 @@ module Database.Esqueleto.Experimental
     , OrderBy
     , DistinctOn
     , LockingKind(..)
+    , LockableEntity(..)
     , SqlString
 
       -- ** Joins
@@ -184,6 +187,10 @@ module Database.Esqueleto.Experimental
     , FullOuterJoin(..)
     , JoinKind(..)
     , OnClauseWithoutMatchingJoinException(..)
+      -- *** Join Helpers
+    , getTable
+    , getTableMaybe
+    , GetFirstTable(..)
 
       -- ** SQL backend
     , SqlQuery
@@ -334,8 +341,8 @@ import Database.Esqueleto.Experimental.ToMaybe
 -- @
 -- select $
 -- from $ \\(people \`LeftOuterJoin\` blogPosts) -> do
--- on (people ^. PersonId ==. blogPosts ?. BlogPostAuthorId)
--- where_ (people ^. PersonAge >. val 18)
+-- on (just (people ^. PersonId) ==. blogPosts ?. BlogPostAuthorId)
+-- where_ (people ^. PersonAge >. just (val 18))
 -- pure (people, blogPosts)
 -- @
 --
@@ -349,8 +356,8 @@ import Database.Esqueleto.Experimental.ToMaybe
 --     from $ table \@Person
 --     \`leftJoin\` table \@BlogPost
 --     \`on\` (\\(people :& blogPosts) ->
---             people ^. PersonId ==. blogPosts ?. BlogPostAuthorId)
--- where_ (people ^. PersonAge >. val 18)
+--             just (people ^. PersonId) ==. blogPosts ?. BlogPostAuthorId)
+-- where_ (people ^. PersonAge >. just (val 18))
 -- pure (people, blogPosts)
 -- @
 --